Add float64 type cast for mean,mean_var
Add float64 type cast for NumericalColumn mean, mean_var,
it will be required if input dtype is not float64.

+++ python/cudf/dataframe/numerical.py
@@ -246,12 +246,12 @@ class NumericalColumn(columnops.TypedColumnBase):
         return cpp_reduce.apply_reduce('product', self, dtype=dtype)
 
     def mean(self, dtype=np.float64):
-        return self.sum(dtype=dtype) / self.valid_count
+        return np.float64(self.sum(dtype=dtype)) / self.valid_count
 
     def mean_var(self, ddof=1, dtype=np.float64):
         mu = self.mean(dtype=dtype)
         n = self.valid_count
-        asum = self.sum_of_squares(dtype=dtype)
+        asum = np.float64(self.sum_of_squares(dtype=dtype))
         div = n - ddof
         var = asum / div - (mu ** 2) * n / div
         return mu, var
